refactor: restructure it according to MVVN
