+++ app/src/main/java/app/revanced/manager/compose/ManagerApplication.kt
@@ -20,7 +20,6 @@ class ManagerApplication : Application() {
                 serviceModule,
                 workerModule,
                 viewModelModule,
-                patcherModule,
             )
         }
     }
