+++ app/src/main/java/app/revanced/manager/compose/patcher/worker/PatcherWorker.kt
@@ -1,10 +1,12 @@
-package app.revanced.manager.compose.patcher
+package app.revanced.manager.compose.patcher.worker
 
 import android.content.Context
 import android.util.Log
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import androidx.work.workDataOf
+import app.revanced.manager.compose.patcher.data.repository.PatchesRepository
+import app.revanced.manager.compose.patcher.Session
 import app.revanced.manager.compose.patcher.aapt.Aapt
 import app.revanced.patcher.extensions.PatchExtensions.patchName
 import kotlinx.serialization.Serializable
@@ -15,8 +17,8 @@ import org.koin.core.component.inject
 import java.io.File
 import java.io.FileNotFoundException
 
-class ReVancedWorker(context: Context, parameters: WorkerParameters) : CoroutineWorker(context, parameters), KoinComponent {
-    private val patcherState: PatcherState by inject()
+class PatcherWorker(context: Context, parameters: WorkerParameters) : CoroutineWorker(context, parameters), KoinComponent {
+    private val patcherState: PatchesRepository by inject()
 
     companion object {
         const val Progress = "progress"
